Cambios en las entidades y los servicios#

+++ MonedaConvert/Models/Dtos/CreateAndUpdateCurrencyDto.cs
@@ -5,7 +5,7 @@
         public int? Id { get; set; } 
         public string? Symbol { get; set; }    
         public string? Legend { get; set; } 
-        public decimal ? IC { get; set; }
+        public float ? IC { get; set; }
 
     }
 }
